requirements.txt
aws-cdk-lib
constructs
strands-sdk
boto3
requests
python-dotenv
